fix win 64 build (only page_encryption files)

+++ include/my_aes.h
@@ -54,8 +54,8 @@ C_MODE_START
     != 0           error
     0             no error
 */
-int my_aes_encrypt_cbc(const char* source, unsigned long int source_length,
-					char* dest, unsigned long int *dest_length,
+int my_aes_encrypt_cbc(const char* source, uint32 source_length,
+					char* dest, uint32 *dest_length,
 					const unsigned char* key, uint8 key_length,
 					const unsigned char* iv, uint8 iv_length,
 					int noPadding);
@@ -118,8 +118,8 @@ int my_aes_encrypt(const char *source, int source_length, char *dest,
     != 0           error
     0             no error
 */
-int my_aes_decrypt_cbc(const char* source, unsigned long int source_length,
-					char* dest, unsigned long int *dest_length,
+int my_aes_decrypt_cbc(const char* source, uint32 source_length,
+					char* dest, uint32 *dest_length,
 					const unsigned char* key, uint8 key_length,
 					const unsigned char* iv, uint8 iv_length,
 					int noPadding);
